+++ gui/extension/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixes
 
 - BUG#37498176 Type checking errors for REST procedure without output parameters
+- BUG#37549793 Unexpected MRS authentication error in the Python SDK
 -
 
 ## Changes in 1.19.1+9.1.0